Embedded scroll view for profile creation
This is a simple app to display the functionality of the
EmbeddedScrollViewController class.  The controller responds to button
events to scroll various subviews of the scrollView into the visible
region.  This allows the user to easily work through a series of
related views (a profile creation process, for example) without pushing
a bunch of new view controllers.  Some of the work is done by the
ScrollContentContainer, which has an array of subviews that it lays out
in relation to its parent scrollView.